+++ FmuImporter/FmuImporter/CommDescription/StructMemberInternal.cs
@@ -1,15 +1,38 @@
 ﻿// SPDX-License-Identifier: MIT
 // Copyright (c) Vector Informatik GmbH. All rights reserved.
 
+using FmuImporter.Helpers;
+
 namespace FmuImporter.CommDescription;
 
 public class StructMemberInternal : StructMember
 {
-  public StructMemberInternal(StructMember member)
+  public StructMemberInternal()
+  {
+  }
+
+  public StructMemberInternal(StructMemberInternal member)
   {
     Name = member.Name;
     Type = member.Type;
+    ResolvedType = member.ResolvedType;
+    QualifiedName = member.QualifiedName;
+  }
+
+  public new string Type
+  {
+    get
+    {
+      return base.Type;
+    }
+    set
+    {
+      base.Type = value;
+      ResolvedType = Helpers.Helpers.StringToType(Type);
+    }
   }
 
   public string QualifiedName { get; set; } = default!;
+
+  public OptionalType ResolvedType { get; set; } = default!;
 }